+++ NEWS.md
@@ -1,3 +1,5 @@
+# ggmagnify (development version)
+
 
 # ggmagnify 0.3.0