+++ NEWS.md
@@ -1,4 +1,4 @@
-# ggmagnify (development version)
+# ggmagnify 0.4.0
 
 * New `proj.fill` argument to fill the area between the projection lines.